MAINT: split `pixi` sub-module
