+++ src/compwa_policy/check_dev_files/pixi/_helpers.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from compwa_policy.utilities import CONFIG_PATH
+from compwa_policy.utilities.match import filter_files
+from compwa_policy.utilities.pyproject import Pyproject
+
+
+def has_pixi_config(pyproject: Pyproject | None = None) -> bool:
+    if filter_files(["pixi.lock", "pixi.toml"]):
+        return True
+    if pyproject is not None:
+        return pyproject.has_table("tool.pixi")
+    return CONFIG_PATH.pyproject.exists() and Pyproject.load().has_table("tool.pixi")
